Move context menu stuff lower level component
