+++ webview/src/experiments/components/table/Table.test.tsx
@@ -18,9 +18,9 @@ import { MessageFromWebviewType } from 'dvc/src/webview/contract'
 import React from 'react'
 import { TableInstance } from 'react-table'
 import tableDataFixture from 'dvc/src/test/fixtures/expShow/base/tableData'
-import { SortOrder } from './header/TableHeader'
 import { Table } from './Table'
 import styles from './styles.module.scss'
+import { SortOrder } from './header/ContextMenuContent'
 import { ExperimentsTable } from '../Experiments'
 import * as ColumnOrder from '../../hooks/useColumnOrder'
 import { vsCodeApi } from '../../../shared/api'
