+++ webview/src/experiments/components/table/header/TableHeaderCellContents.tsx
@@ -2,7 +2,7 @@ import React, { useState, useEffect } from 'react'
 import cx from 'classnames'
 import { ColumnType, Experiment } from 'dvc/src/experiments/webview/contract'
 import { HeaderGroup } from 'react-table'
-import { SortOrder } from './TableHeader'
+import { SortOrder } from './ContextMenuContent'
 import styles from '../styles.module.scss'
 import {
   Draggable,
